fix: resolve issue #191 — move DriverRepository calls from Form1 into RaceController
